requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.0
claude-agent-sdk>=0.1.33
